rename all standalone 'exports' resources to 'downloads'
